[PAM-1392] Preserve reportee on updates of approved ads
Administration status is only mapped in pam-ad when a reportee is set.

Ensure reportee is kept on ads in GODKJENT state, when transitioning
to OPPDATERT for ad updates. This in turns causes correct
administration state to be propagated to pam-ad (keeping the status
DONE).

Also, when system marks ads as duplicates, set reportee to "System",
for the same reason as above.